+++ integrations/github/github/core/options.py
@@ -1,9 +1,7 @@
-from typing import NotRequired, Required, TypedDict, List, Optional
+from typing import List, NotRequired, Optional, Required, TypedDict
 
 
 class SingleRepositoryOptions(TypedDict):
-    """Options for fetching a single repository."""
-
     name: str
 
 
@@ -11,6 +9,36 @@ class ListRepositoryOptions(TypedDict):
     """Options for listing repositories."""
 
     type: str
+
+
+class RepositoryIdentifier(TypedDict):
+    """Options for identifying a repository."""
+
+    repo_name: Required[str]
+
+
+class SinglePullRequestOptions(RepositoryIdentifier):
+    """Options for fetching a single pull request."""
+
+    pr_number: Required[int]
+
+
+class ListPullRequestOptions(RepositoryIdentifier):
+    """Options for listing pull requests."""
+
+    state: Required[str]
+
+
+class SingleIssueOptions(RepositoryIdentifier):
+    """Options for fetching a single issue."""
+
+    issue_number: Required[int]
+
+
+class ListIssueOptions(RepositoryIdentifier):
+    """Options for listing issues."""
+
+    state: Required[str]
 
 
 class FileSearchOptions(TypedDict):